Implement openEuler scripted install
openEuler 24.03 seems closest to el8, so start with that baseline.

Modify el8 handling to recognize a totally missing CA bundle and assume
the original el8 location.

Have osimage recognize openEuler media by presence of openEuler release file.

+++ confluent_osdeploy/el8/initramfs/usr/lib/dracut/hooks/cmdline/01-confluent.sh
@@ -3,10 +3,12 @@ echo -n "" >> /tmp/net.ifaces
 echo -n "" > /tmp/01-autocons.devnode
 BUNDLENAME=/etc/pki/tls/certs/ca-bundle.crt
 if [ ! -e "$BUNDLENAME" ]; then
-    BUNDLENAME=/etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem
     mkdir -p /etc/pki/tls/certs
-    ln -s $BUNDLENAME /etc/pki/tls/certs/ca-bundle.crt
-fi
+    if [ -e /etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem ]; then
+        BUNDLENAME=/etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem
+        ln -s $BUNDLENAME /etc/pki/tls/certs/ca-bundle.crt
+    fi
+fi	
 cat /tls/*.0 >> $BUNDLENAME
 if ! grep console= /proc/cmdline >& /dev/null; then
     autocons=$(/opt/confluent/bin/autocons)
